Restructure connected ecosystems components with improved interactions
